Jellyfish-buffer refactor for separation of BufferComposer from jellyfish-transaction
