+++ .eslintrc
@@ -20,7 +20,10 @@
       "files": [
         "**/jellyfish-transaction/__tests__/**/*.test.ts",
         "**/jellyfish-transaction/src/script/dftx/**/*.ts",
-        "**/jellyfish-transaction/src/*.ts"
+        "**/jellyfish-transaction/src/*.ts",
+        "**/jellyfish-buffer/__tests__/**/*.test.ts",
+        "**/jellyfish-buffer/src/script/dftx/**/*.ts",
+        "**/jellyfish-buffer/src/*.ts"
       ],
       "rules": {
         "no-return-assign": "off"
