+++ .github/labeler.yml
@@ -31,6 +31,11 @@ labels:
     matcher:
       files: "packages/jellyfish-api-jsonrpc/**"
 
+  - label: area/jellyfish-buffer
+    sync: true
+    matcher:
+      files: "packages/jellyfish-buffer/**"
+
   - label: area/jellyfish-crypto
     sync: true
     matcher:
