+++ .github/labels.yml
@@ -65,6 +65,8 @@
 - color: fbca04
   name: area/jellyfish-api-jsonrpc
 - color: fbca04
+  name: area/jellyfish-buffer
+- color: fbca04
   name: area/jellyfish-crypto
 - color: fbca04
   name: area/jellyfish-json
