+++ packages/jellyfish-buffer/README.md
@@ -1,0 +1,5 @@
+[![npm](https://img.shields.io/npm/v/@defichain/jellyfish-buffer)](https://www.npmjs.com/package/@defichain/jellyfish-buffer/v/latest)
+
+# @defichain/jellyfish-buffer
+
+DeFi Blockchain buffer composer.
