+++ packages/jellyfish-buffer/src/bitmask.ts
