+++ packages/jellyfish-transaction/src/script/dftx/dftx_icxorderbook.ts
@@ -1,4 +1,4 @@
-import { BufferComposer, ComposableBuffer } from '../../buffer/buffer_composer'
+import { BufferComposer, ComposableBuffer } from '@defichain/jellyfish-buffer'
 import BigNumber from 'bignumber.js'
 import { Script } from '../../tx'
 import { CScript } from '../../tx_composer'
