+++ packages/jellyfish-transaction/src/script/dftx/dftx_masternode.ts
@@ -1,5 +1,5 @@
 import { SmartBuffer } from 'smart-buffer'
-import { BufferComposer, ComposableBuffer } from '../../buffer/buffer_composer'
+import { BufferComposer, ComposableBuffer } from '@defichain/jellyfish-buffer'
 
 /**
  * CreateMasternode DeFi Transaction
