+++ packages/jellyfish-transaction/src/script/dftx/dftx_misc.ts
@@ -1,4 +1,4 @@
-import { BufferComposer, ComposableBuffer } from '../../buffer/buffer_composer'
+import { BufferComposer, ComposableBuffer } from '@defichain/jellyfish-buffer'
 
 /**
  * Composable UtxosToAccount, C stands for Composable.
