+++ packages/jellyfish-transaction/src/script/dftx/dftx_oracles.ts
@@ -1,4 +1,4 @@
-import { BufferComposer, ComposableBuffer } from '../../buffer/buffer_composer'
+import { BufferComposer, ComposableBuffer } from '@defichain/jellyfish-buffer'
 import { CCurrencyPair, CTokenPrice, CurrencyPair, TokenPrice } from './dftx_price'
 import { Script } from '../../tx'
 import { CScript } from '../../tx_composer'
