+++ packages/jellyfish-transaction/src/tx_composer.ts
@@ -1,9 +1,8 @@
 import BigNumber from 'bignumber.js'
 import { SmartBuffer } from 'smart-buffer'
-import { BufferComposer, ComposableBuffer } from './buffer/buffer_composer'
+import { BufferComposer, ComposableBuffer, readVarUInt, writeVarUInt } from '@defichain/jellyfish-buffer'
 import { Script, Transaction, TransactionSegWit, Vin, Vout, Witness, WitnessScript } from './tx'
 import { OP_CODES, OPCode } from './script'
-import { readVarUInt, writeVarUInt } from './buffer/buffer_varuint'
 import { dSHA256 } from '@defichain/jellyfish-crypto'
 
 /**
