+++ tsconfig.json
@@ -6,6 +6,7 @@
       "@defichain/jellyfish-address": ["jellyfish-address/src"],
       "@defichain/jellyfish-api-core": ["jellyfish-api-core/src"],
       "@defichain/jellyfish-api-jsonrpc": ["jellyfish-api-jsonrpc/src"],
+      "@defichain/jellyfish-buffer": ["jellyfish-bufferc/src"],
       "@defichain/jellyfish-crypto": ["jellyfish-crypto/src"],
       "@defichain/jellyfish-json": ["jellyfish-json/src"],
       "@defichain/jellyfish-network": ["jellyfish-network/src"],
@@ -23,6 +24,7 @@
       "@defichain/jellyfish-address/*": ["jellyfish-address/src/*"],
       "@defichain/jellyfish-api-core/*": ["jellyfish-api-core/src/*"],
       "@defichain/jellyfish-api-jsonrpc/*": ["jellyfish-api-jsonrpc/src/*"],
+      "@defichain/jellyfish-buffer/*": ["jellyfish-bufferc/src/*"],
       "@defichain/jellyfish-crypto/*": ["jellyfish-crypto/src/*"],
       "@defichain/jellyfish-json/*": ["jellyfish-json/src/*"],
       "@defichain/jellyfish-network/*": ["jellyfish-network/src/*"],
